+++ prequ/scripts/compile.py
@@ -32,6 +32,7 @@ class PipCommand(pip.basecommand.Command):
 @click.command()
 @click.version_option()
 @click.option('-v', '--verbose', is_flag=True, help="Show more output")
+@click.option('-s', '--silent', is_flag=True, help="Show no output")
 @click.option('-n', '--dry-run', is_flag=True, help="Only show what would happen, don't change anything")
 @click.option('-p', '--pre', is_flag=True, default=None, help="Allow resolving to prereleases (default is not)")
 @click.option('-r', '--rebuild', is_flag=True, help="Clear any caches upfront, rebuild from scratch")
@@ -62,10 +63,10 @@ class PipCommand(pip.basecommand.Command):
 @click.option('--generate-hashes', is_flag=True, default=False,
               help="Generate pip 8 style hashes in the resulting requirements file.")
 @click.argument('src_files', nargs=-1, type=click.Path(exists=True, allow_dash=True))
-def cli(verbose, dry_run, pre, rebuild, find_links, index_url, extra_index_url,
-        client_cert, trusted_host, header, index, emit_trusted_host, annotate,
-        upgrade, upgrade_packages, output_file, allow_unsafe, generate_hashes,
-        src_files):
+def cli(verbose, silent, dry_run, pre, rebuild, find_links, index_url,
+        extra_index_url, client_cert, trusted_host, header, index,
+        emit_trusted_host, annotate, upgrade, upgrade_packages, output_file,
+        allow_unsafe, generate_hashes, src_files):
     """Compiles requirements.txt from requirements.in specs."""
     log.verbose = verbose
 
@@ -239,7 +240,8 @@ def cli(verbose, dry_run, pre, rebuild, find_links, index_url, extra_index_url,
                           trusted_hosts=pip_options.trusted_hosts,
                           format_control=repository.finder.format_control,
                           find_links=repository.finder.find_links,
-                          allow_unsafe=allow_unsafe)
+                          allow_unsafe=allow_unsafe,
+                          silent=silent)
     writer.write(results=results,
                  reverse_dependencies=reverse_dependencies,
                  primary_packages=primary_packages,