+++ ChangeLog.rst
@@ -2,6 +2,7 @@ Unreleased
 ----------
 
 - (nvie/pip-tools#453) Write relative find-links opts to output file
+- Add "--silent" option for the compile command
 
 0.180.8
 -------